Re-add add order functionality

+++ src/main/java/web/javaproject/fooddeliveryapp/controller/OrderController.java
@@ -1,7 +1,12 @@
 package web.javaproject.fooddeliveryapp.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -11,10 +16,14 @@ import web.javaproject.fooddeliveryapp.dto.UpdateOrderDTO;
 import web.javaproject.fooddeliveryapp.mapper.OrderMapper;
 import web.javaproject.fooddeliveryapp.model.Client;
 import web.javaproject.fooddeliveryapp.model.Order;
+import web.javaproject.fooddeliveryapp.model.security.Authority;
+import web.javaproject.fooddeliveryapp.model.security.CustomUserDetails;
 import web.javaproject.fooddeliveryapp.service.ClientService;
 import web.javaproject.fooddeliveryapp.service.OrderService;
+import java.util.Arrays;
 
 import java.util.List;
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/order")
@@ -34,14 +43,37 @@ public class OrderController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<?> create(@RequestBody CreateOrderDTO createOrderDTO) {
+    public String create(HttpServletRequest request, Model model) {
         try {
+            Long restaurantId = Long.parseLong(request.getParameter("restaurantId"));
+            Long clientId = Long.parseLong(request.getParameter("clientId"));
+            List<Long> dishesIds = Arrays.stream(request.getParameter("dishesIds").split(","))
+                    .map(Long::valueOf)
+                    .toList();
+
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+            GrantedAuthority firstAuthority = authentication.getAuthorities().stream().findFirst().orElse(null);
+            if(firstAuthority == null || !Objects.equals(firstAuthority.getAuthority(), "CLIENT")) {
+                return "access_denied";
+            }
+
+            CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+            if(!Objects.equals(userDetails.getAssociatedId(), clientId)) {
+                return "access_denied";
+            }
+
+            CreateOrderDTO createOrderDTO = new CreateOrderDTO(clientId, restaurantId, dishesIds);
+
             Order order = orderService.createOrder(createOrderDTO);
             OrderDTO orderDTO = orderMapper.toDTO(order);
 
-            return new ResponseEntity<>(orderDTO, HttpStatus.CREATED);
+            model.addAttribute("order", orderDTO);
+            model.addAttribute("orderId", order.getId());
+            return "orderView";
         } catch (Exception e) {
-            return new ResponseEntity<>("Error creating order: " + e.getMessage(), HttpStatus.BAD_REQUEST);
+            model.addAttribute(e.getMessage(), "errorMessage");
+            return "errorView";
         }
     }
 
@@ -50,6 +82,29 @@ public class OrderController {
         try {
             Order order = orderService.getOrder(orderId);
             OrderDTO orderDTO = orderMapper.toDTO(order);
+
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            GrantedAuthority firstAuthority = authentication.getAuthorities().stream().findFirst().orElse(null);
+            if(firstAuthority == null) {
+                return "access_denied";
+            }
+
+            CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+
+            switch (firstAuthority.getAuthority()) {
+                case "CLIENT":
+                    if(!Objects.equals(userDetails.getAssociatedId(), order.getClient().getId())) {
+                        return "access_denied";
+                    }
+                case "RESTAURANT":
+                    if(!Objects.equals(userDetails.getAssociatedId(), order.getRestaurant().getId())) {
+                        return "access_denied";
+                    }
+                case "COURIER":
+                    if(!Objects.equals(userDetails.getAssociatedId(), order.getCourier().getId())) {
+                        return "access_denied";
+                    }
+            }
 
             model.addAttribute("order", orderDTO);
             model.addAttribute("orderId", order.getId());
@@ -80,36 +135,45 @@ public class OrderController {
     }
 
     @PutMapping("/{orderId}")
-    public ResponseEntity<?> updateOrder(@PathVariable Long orderId, @RequestBody UpdateOrderDTO updateOrderDTO) {
+    public String updateOrder(@PathVariable Long orderId, @RequestBody UpdateOrderDTO updateOrderDTO, Model model) {
         try {
             Order updatedOrder = orderService.updateOrder(orderId, updateOrderDTO);
             OrderDTO updatedOrderDTO = orderMapper.toDTO(updatedOrder);
 
-            return new ResponseEntity<>(updatedOrderDTO, HttpStatus.OK);
+            model.addAttribute("order", updatedOrderDTO);
+            model.addAttribute("orderId", updatedOrder.getId());
+
+            return "orderView";
         } catch (Exception e) {
-            return new ResponseEntity<>("Error updating order: " + e.getMessage(), HttpStatus.BAD_REQUEST);
+            model.addAttribute("errorMessage", e.getMessage());
+            return "errorView";
         }
     }
 
     @DeleteMapping("/{orderId}")
-    public ResponseEntity<?> deleteOrder(@PathVariable Long orderId) {
+    public String deleteOrder(@PathVariable Long orderId, Model model) {
         try {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            GrantedAuthority firstAuthority = authentication.getAuthorities().stream().findFirst().orElse(null);
+            if(firstAuthority == null) {
+                return "access_denied";
+            }
+
+            Order order = orderService.getOrder(orderId);
+            if(Objects.equals(firstAuthority.getAuthority(), "CLIENT")) {
+                CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+
+                if(!Objects.equals(order.getClient().getId(), userDetails.getAssociatedId())) {
+                    return "acess_denied";
+                }
+            }
+
             orderService.deleteOrder(orderId);
-            return new ResponseEntity<>("Order deleted successfully", HttpStatus.OK);
+            return "redirect:/order";
         } catch (Exception e) {
-            return new ResponseEntity<>("Error deleting order: " + e.getMessage(), HttpStatus.BAD_REQUEST);
+            model.addAttribute("errorMessage", e.getMessage());
+            return "errorView";
         }
     }
 }
-//
-//    @PostMapping("/create")
-//    public ResponseEntity<?> create(@RequestBody CreateOrderDTO createOrderDTO) {
-//        try {
-//            Order order = orderService.createOrder(createOrderDTO);
-//            OrderDTO orderDTO = orderMapper.toDTO(order);
-//
-//            return new ResponseEntity<>(orderDTO, HttpStatus.CREATED);
-//        } catch (Exception e) {
-//            return new ResponseEntity<>("Error creating order: " + e.getMessage(), HttpStatus.BAD_REQUEST);
-//        }
-//    }
+
